+++ src/main/java/controller/LottoController.java
@@ -3,8 +3,8 @@ package controller;
 import static constant.ExceptionMessages.INVALID_WINNING_NUMBERS_SIZE_EXCEPTION_MESSAGE;
 import static constant.LottoConstants.LOTTO_PRICE;
 import static constant.LottoConstants.WINNING_NUMBERS_SIZE;
-import static validator.LottoNumberValidators.validateNoDuplicateInList;
-import static validator.LottoNumberValidators.validateNoDuplicates;
+import static validator.NumberValidators.validateNoDuplicateInList;
+import static validator.NumberValidators.validateNoDuplicates;
 
 import domain.LottoGame;
 import domain.LottoNumber;
@@ -13,7 +13,7 @@ import domain.Lottos;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import validator.LottoNumberValidators;
+import validator.NumberValidators;
 import view.InputView;
 import view.OutputView;
 
@@ -22,7 +22,9 @@ public class LottoController {
     public void run() {
         Lottos lottos = initCustomerLottos();
         LottoReferee referee = initLottoReferee();
+
         LottoGame lottoGame = new LottoGame(lottos, referee);
+
         OutputView.printLottoResults(lottoGame.getResultStatistics());
         OutputView.printLottoResults(lottoGame.calculateProfitRatio());
     }
@@ -44,7 +46,7 @@ public class LottoController {
     private List<LottoNumber> registerWinningNumbers() {
         String winningNumbersInput = InputView.requestWinningNumbers();
         List<LottoNumber> winningNumbers = Arrays.stream(winningNumbersInput.split(", "))
-                .map(LottoNumberValidators::validateAndParseNumber)
+                .map(NumberValidators::validateAndParseNumber)
                 .map(LottoNumber::of)
                 .collect(Collectors.toList());
         if (winningNumbers.size() != WINNING_NUMBERS_SIZE) {
